[Ilham] Rename package to 'suit'
