+++ app/build.gradle
@@ -8,7 +8,7 @@ android {
     buildToolsVersion "30.0.3"
 
     defaultConfig {
-        applicationId "com.rockpaperscissors.team.b"
+        applicationId "com.suit.team.b"
         minSdkVersion 16
         targetSdkVersion 30
         versionCode 1
