--- app/src/main/java/com/rockpaperscissors/team/b/App.kt
@@ -1,15 +1,0 @@
-package com.rockpaperscissors.team.b
-
-import android.app.Application
-import android.content.Context
-
-class App : Application() {
-    companion object {
-        var context: Context? = null
-    }
-
-    override fun onCreate() {
-        super.onCreate()
-        context = applicationContext
-    }
-}
