--- app/src/main/java/com/rockpaperscissors/team/b/ui/profile/show/ProfilePresenter.kt
@@ -1,6 +1,0 @@
-package com.rockpaperscissors.team.b.ui.profile.show
-
-interface ProfilePresenter {
-    fun update()
-    fun delete()
-}
