--- app/src/main/java/com/rockpaperscissors/team/b/ui/profile/show/ProfileView.kt
@@ -1,6 +1,0 @@
-package com.rockpaperscissors.team.b.ui.profile.show
-
-interface ProfileView {
-    fun onSuccess(toastString: String)
-    fun onFailed(toastString: String)
-}
