--- app/src/main/java/com/rockpaperscissors/team/b/ui/score/ScorePresenter.kt
@@ -1,6 +1,0 @@
-package com.rockpaperscissors.team.b.ui.score
-
-interface ScorePresenter {
-    fun getRankVsP()
-    fun getRankVsCPU()
-}
