--- app/src/main/java/com/rockpaperscissors/team/b/ui/score/ScoreView.kt
@@ -1,7 +1,0 @@
-package com.rockpaperscissors.team.b.ui.score
-
-import com.rockpaperscissors.team.b.data.model.Score
-
-interface ScoreView {
-    fun onSuccess(dataScore: MutableList<Score>)
-}
