--- app/src/main/java/com/rockpaperscissors/team/b/utils/GameType.kt
@@ -1,6 +1,0 @@
-package com.rockpaperscissors.team.b.utils
-
-enum class GameType {
-    VSP,
-    VSCPU
-}
