--- app/src/main/java/com/rockpaperscissors/team/b/utils/PlayerType.kt
@@ -1,7 +1,0 @@
-package com.rockpaperscissors.team.b.utils
-
-enum class PlayerType {
-    P1,
-    P2,
-    CPU
-}
