+++ app/src/main/java/com/suit/team/b/App.kt
@@ -1,0 +1,15 @@
+package com.suit.team.b
+
+import android.app.Application
+import android.content.Context
+
+class App : Application() {
+    companion object {
+        var context: Context? = null
+    }
+
+    override fun onCreate() {
+        super.onCreate()
+        context = applicationContext
+    }
+}
