+++ app/src/main/java/com/suit/team/b/ui/profile/show/ProfilePresenter.kt
@@ -1,0 +1,6 @@
+package com.suit.team.b.ui.profile.show
+
+interface ProfilePresenter {
+    fun update()
+    fun delete()
+}
