+++ app/src/main/java/com/suit/team/b/ui/profile/show/ProfileView.kt
@@ -1,0 +1,6 @@
+package com.suit.team.b.ui.profile.show
+
+interface ProfileView {
+    fun onSuccess(toastString: String)
+    fun onFailed(toastString: String)
+}
