+++ app/src/main/java/com/suit/team/b/ui/score/ScorePresenter.kt
@@ -1,0 +1,6 @@
+package com.suit.team.b.ui.score
+
+interface ScorePresenter {
+    fun getRankVsP()
+    fun getRankVsCPU()
+}
