+++ app/src/main/java/com/suit/team/b/ui/score/ScorePresenterImp.kt
@@ -1,0 +1,17 @@
+package com.suit.team.b.ui.score
+
+import com.suit.team.b.data.local.SharedPref
+
+class ScorePresenterImp(private val view: ScoreView) : ScorePresenter {
+
+    override fun getRankVsP() {
+        val dataScoreVsP = SharedPref.getRankedScoreVsP()
+        view.onSuccess(dataScoreVsP)
+    }
+
+    override fun getRankVsCPU() {
+        val dataScoreVsCom = SharedPref.getRankedScoreVsCPU()
+        view.onSuccess(dataScoreVsCom)
+    }
+
+}
