+++ app/src/main/java/com/suit/team/b/ui/score/ScoreView.kt
@@ -1,0 +1,7 @@
+package com.suit.team.b.ui.score
+
+import com.suit.team.b.data.model.Score
+
+interface ScoreView {
+    fun onSuccess(dataScore: MutableList<Score>)
+}
