+++ app/src/main/java/com/suit/team/b/utils/GameType.kt
@@ -1,0 +1,6 @@
+package com.suit.team.b.utils
+
+enum class GameType {
+    VSP,
+    VSCPU
+}
