+++ app/src/main/java/com/suit/team/b/utils/PlayerType.kt
@@ -1,0 +1,7 @@
+package com.suit.team.b.utils
+
+enum class PlayerType {
+    P1,
+    P2,
+    CPU
+}
