+++ app/src/main/res/layout/dummy_activity.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ui.dummy.DummyActivity"
+    tools:context="com.suit.team.b.ui.dummy.DummyActivity"
     >
 
     <EditText
