+++ app/src/main/res/layout/profilepage_activity.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".ui.profile.show.ProfilePageActivity">
+    tools:context="com.suit.team.b.ui.profile.show.ProfilePageActivity">
 
     <TextView
         android:id="@+id/tvUsername"
